Ajuste do no namespace das classes referente ao banco de dados. Inclusão das classes de integração de produto, empresa e transportadora.

+++ FWLog.Data/Models/EmpresaTipo.cs
@@ -5,7 +5,7 @@ namespace FWLog.Data.Models
     public class EmpresaTipo
     {
         [Key]
-        public int IdEmpresaTipo { get; set; }
+        public long IdEmpresaTipo { get; set; }
 
         [StringLength(50)]
         public string Descricao { get; set; }
